+++ src/main/java/lotto/RandomNumber.java
@@ -1,29 +1,7 @@
 package lotto;
 
-import camp.nextstep.edu.missionutils.Randoms;
+import java.util.List;
 
-import java.util.*;
-
-import static lotto.Constant.*;
-
-
-public class RandomNumber {
-    private static RandomNumber randomNumber;
-
-    private RandomNumber() {
-    }
-
-    public static RandomNumber getInstance() {
-        if (randomNumber == null) {
-            randomNumber = new RandomNumber();
-        }
-        return randomNumber;
-    }
-
-    public List<Integer> getRandomNumber() {
-        List<Integer> randomNumber = new ArrayList<>(Randoms.pickUniqueNumbersInRange(MIN_BOUND, MAX_BOUND, DIGITS));
-        Collections.sort(randomNumber);
-        System.out.println(randomNumber);
-        return randomNumber;
-    }
+public interface RandomNumber {
+    public List<Integer> getRandomNumber();
 }
